reorganized test constants
