+++ tests/tests/constants/testContracts.ts
@@ -1,23 +1,4 @@
-import { SignedTransaction, TransactionConfig } from "web3-core";
 import { AbiItem } from "web3-utils";
-
-export const PORT = 19931;
-export const RPC_PORT = 19932;
-export const WS_PORT = 19933;
-export const SPECS_PATH = `./moonbeam-test-specs`;
-
-export const DISPLAY_LOG = process.env.MOONBEAM_LOG || false;
-export const MOONBEAM_LOG = process.env.MOONBEAM_LOG || "info";
-
-export const BINARY_PATH =
-  process.env.BINARY_PATH || `../node/standalone/target/release/moonbase-standalone`;
-export const SPAWNING_TIME = 30000;
-
-// Test variables
-export const GENESIS_ACCOUNT = "0x6be02d1d3665660d22ff9624b7be0551ee1ac91b";
-export const GENESIS_ACCOUNT_PRIVATE_KEY =
-  "0x99B3C12287537E38C90A9219D4CB074A89A16E9CDB20BF85728EBD97C343E342";
-export const TEST_ACCOUNT = "0x1111111111111111111111111111111111111111";
 
 // Solidity: contract test {function multiply(uint a) public pure returns(uint d) {return a * 7;}}
 export const TEST_CONTRACT_BYTECODE =
@@ -131,27 +112,3 @@ export const INFINITE_CONTRACT_ABI_VAR = [
     type: "function",
   },
 ] as AbiItem[];
-
-// +++ TransactionConfig +++
-
-export const basicTransfertx: TransactionConfig = {
-  from: GENESIS_ACCOUNT,
-  to: TEST_ACCOUNT,
-  value: "0x200", // =512 Must me higher than ExistentialDeposit (500)
-  gasPrice: "0x01",
-  gas: "0x100000",
-};
-export const contractCreation: TransactionConfig = {
-  from: GENESIS_ACCOUNT,
-  data: TEST_CONTRACT_BYTECODE,
-  value: "0x00",
-  gasPrice: "0x01",
-  gas: "0x100000",
-};
-
-
-// TESTING NOTES
-// - block dont seem to have gas limit but it's usually around 1500 tx per block 
-// and the time it takes to construct the block increases until around 12s for 1500tx
-// - after the first 1500tx block, following block have aroudn 100-300 tx per block until all blocks are incuded. 10 blockds for 3000tx
-// - infinite loop contract should throw out of gas error, but they don't and they are included in the block. there are some rpc errors sometimes
